extracts the dashboard routes
